requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
mysql-connector==2.2.9
requests==2.32.3
urllib3==2.2.2
persiantools==4.0.3
